+++ python_ta/checkers/shadowing_in_comprehension_checker.py
@@ -26,7 +26,7 @@ class ShadowingInComprehensionChecker(BaseChecker):
                 if target.name in node.parent.frame().locals and target.name != "_":
                     args = target.name
                     self.add_message("shadowing-in-comprehension", node=target, args=args)
-        else:  # isinstance(node.target, nodes.AssignName)
+        elif isinstance(node.target, nodes.AssignName):
             if node.target.name in node.parent.frame().locals and node.target.name != "_":
                 args = node.target.name
                 self.add_message("shadowing-in-comprehension", node=node.target, args=args)